requirements.txt
functions-framework==3.5.0
python-telegram-bot==21.1.1
gspread==5.7.0
oauth2client==4.1.3